Fixed bug with too larg numeric values in json streaming map and list: always using long values, now.

+++ src/main/java/terrastore/util/json/JsonStreamingList.java
@@ -53,7 +53,7 @@ public class JsonStreamingList extends AbstractList {
                     } else if (token.equals(JsonToken.VALUE_NUMBER_FLOAT)) {
                         return parser.getFloatValue();
                     } else if (token.equals(JsonToken.VALUE_NUMBER_INT)) {
-                        return parser.getIntValue();
+                        return parser.getLongValue();
                     } else if (token.equals(JsonToken.VALUE_TRUE) || token.equals(JsonToken.VALUE_FALSE)) {
                         return parser.getBooleanValue();
                     } else {
